Eventos ICPC/MFP, seeds OCI, sync CSES y diseño del juez propio
Standings de eventos internacionales (Controllers/IcpcEvents/):
- Parser BOCA (scorelatam) y parser cphof (World Finals), elegidos por host.
- Endpoints públicos catalog/get + import y edición admin (miembros, país).
- Enriquecimiento contra usuarios registrados: nombre real, @nickname, rating
  y organización (con logo) cuando la fuente no trae institución.
- Seeds: LATAM 2017 y 2019-2025, PdA 2024-2026, WF 2013-2025, MFP fase 2026.
  latam2017 se construyó desde el gym 101889 + listas de icpc.global; el MFP
  desde el standings del grupo de CF (país por bandera, chilenas por amigas).

OCI: seeding también desde CSV (mismo parser que el upload admin) y puntajes
decimales (los ponderados no son enteros).

CSES: el import mantiene un contest por categoría + entrenamiento "cses", y el
sync de resueltos ahora cubre a todos los usuarios (incluidos inactivos).

Standings de contest: incluyen usuarios inactivos (son datos históricos) y
exponen el rating para colorear nombres.

Infra: el job db-reset ya no dropea la base en cada `up`; wwwroot pasa a un
volumen para que los standings subidos sobrevivan rebuilds; seeder incremental.

Docs: diseño del juez propio sobre isolate (un motor, modos judge/run, el
runner del Apunte incluido) y spec implementable del MVP con workers
distribuidos en varios PCs (WSL2) y calibración de velocidad.

+++ ProgCompJOlivaApi/Controllers/Oci/Dtos/OciStandings.cs
@@ -69,11 +69,13 @@ public class OciStandingsRow
     /// <summary>The participant's full real name as printed in the document.</summary>
     public string User { get; set; } = "";
 
-    /// <summary>Per-task scores, aligned with <see cref="OciStandings.Problems"/>.</summary>
-    public List<int> Scores { get; set; } = [];
+    /// <summary>Per-task scores, aligned with <see cref="OciStandings.Problems"/>. May be
+    /// fractional (e.g. the weighted columns of a clasificatoria aggregate).</summary>
+    public List<double> Scores { get; set; } = [];
 
-    /// <summary>Total score: sum of <see cref="Scores"/>, or the weighted total on a weighted aggregate.</summary>
-    public int Global { get; set; }
+    /// <summary>Total score: sum of <see cref="Scores"/>, or the weighted total on a weighted
+    /// aggregate. May be fractional (e.g. 214.7).</summary>
+    public double Global { get; set; }
 
     /// <summary>
     /// Whether the participant advanced to the next stage. Meaning depends on the standings: regional →
